refactor: isolate common memory accounting utils
